[ADD] weblate: Add support for weblate translations
[REF] test_server, travis_weblate: Compatibility with oca/mqt project

[ADD] weblate: Add support for weblate translations

[FIX] travis_weblate: Adding method main to the travis_weblate

[FIX] travis_weblate: Fix the way of calculate self._langs

[IMP] travis_weblate: Register one pull request when exist conflict

[IMP] travis_weblate: Adding mention to the commit author in the conflick

[REF] travis_weblate: Liberate the weblate repository when exist conflict

[REF] travis_weblate: Select the current branch

[FIX] travis_weblate: Merge with squash

[IMP] git_run.py: Debug all git command

[REF] travis_weblate.py: Reset the weblate repository

[FIX] travis_weblate.py: Git fetch all only one commit

[REF] travis_weblate.py: Commit after generate odoo po files

[FIX] travis_weblate.py: Change the order of pull --squash and the generate odoo files for avoid conflict

[FIX] travis_weblate.py: Apply the patch of odoo for export empty string instead the same string

[REF] travis_weblate.py: Adding docstring to the method _apply_patch_odoo

[FIX] travis_install_nightly: Apply weblate path before odoo run

[REF] travis_install_nightly: Print weblate path info

[REF] travis_weblate.py: Delete unnecessary if and add validation when WEBLATE = 1

[FIX] travis_install_nightly: Debug weblate variable

[FIX] travis_install_nightly: Print all environment variable

[FIX] travis_weblate.py: travis_weblate.py: Apply the patch of odoo for export empty string instead the same string

[FIX] travis_weblate.py: Adding validation when no component

[FIX] travis_weblate.py: Reset hard based on the origin

[REF] travis_weblate.py: Created method _check for check the api conection

[REF] apis.py: Raise exception instead print and exit

[REF] weblate: Add support for weblate translations

[FIX] apis.py: Use odoo-community.org in place of vauxoo.com

[FIX] travis_weblate.py: Avoid hardcode name of database

Better use the envinment variable MQT_TEST_DB

+++ travis/test_server.py
@@ -415,7 +415,7 @@ def main(argv=None):
                 counted_errors += errors
                 all_errors.append(to_test)
                 print(fail_msg, "Found %d lines with errors" % errors)
-        if not instance_alive:
+        if not instance_alive and odoo_unittest:
             # Don't drop the database if will be used later.
             subprocess.call(["dropdb", database])
